+++ eio-okta-api/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.2.1
+
+- (DEPENDENCY) eio-okta-data 0.2.0 -> 0.2.1
+
 ## 0.2.0
 
 - (LICENSE) more flexible licensing, changed from `MIT` to `MIT OR Apache-2.0`.