+++ eio-okta-data/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.1
+
+- (COMPATIBILITY) impl CityNameGenFn for fake::locales::Custom
+  - required for compatibility with fake v3.1.0+
+
 ## 0.2.0
 
 - (LICENSE) more flexible licensing, changed from `MIT` to `MIT OR Apache-2.0`.